add wireless to tab and separate config and authentication
