+++ src/lib/y2network/dialogs/edit_interface.rb
@@ -5,6 +5,7 @@ require "y2network/widgets/bond_slaves_tab.rb"
 require "y2network/widgets/bridge_slaves_tab.rb"
 require "y2network/widgets/general_tab.rb"
 require "y2network/widgets/hardware_tab.rb"
+require "y2network/widgets/wireless_tab.rb"
 
 Yast.import "LanItems"
 Yast.import "Label"
@@ -27,9 +28,10 @@ module Y2Network
       end
 
       def contents
-        # if there is addr, make it initial
+        # if there is addr, make it initial unless for wifi, where first one should be wifi specific
+        # configs
         addr_tab = Widgets::AddressTab.new(@settings)
-        addr_tab.initial = true
+        addr_tab.initial = true unless @settings.type.wireless?
 
         tabs = case @settings.type.short_name
         when "vlan"
@@ -41,6 +43,10 @@ module Y2Network
         when "bond"
           [Widgets::GeneralTab.new(@settings), addr_tab, Widgets::HardwareTab.new(@settings),
            Widgets::BondSlavesTab.new(@settings)]
+        when "wlan"
+          wireless = Widgets::WirelessTab.new(@settings)
+          wireless.initial = true
+          [Widgets::GeneralTab.new(@settings), wireless, addr_tab, Widgets::HardwareTab.new(@settings)]
         else
           [Widgets::GeneralTab.new(@settings), addr_tab, Widgets::HardwareTab.new(@settings)]
         end
@@ -56,7 +62,6 @@ module Y2Network
       # removes back button when editing device, but keep it when this dialog follows adding
       # new interface
       def back_button
-        # TODO: decide operation based on builder
         @settings.newly_added? ? Yast::Label.BackButton : ""
       end
     end
